% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{assayize}
\alias{assayize}
\title{add summaries about distance to flanking snps to assay rows}
\usage{
assayize(V, targets, consSeq = NULL, reqDist = 20, allVar = TRUE)
}
\arguments{
\item{V}{a data frame of all the variants detected in sequences from individuals.
Typically this will be from a VCF file (maybe just the first few columns). It can
have any number of columns, but it must have the following:
\describe{
   \item{CHROM}{The name of the contiguous piece of DNA that the variant is found in.
     This will typically be a RAD locus identifier, etc.}
   \item{POS}{The position of this particular variant within the CHROM DNA.  Starts
     counting from 1.}
   \item{REF}{The nucleotide base of the reference sequence at each of the positions.  Must be uppercase A, C, G, or T}
   \item{ALT}{The alternate base(s) at each of the positions.  Must be uppercase, A, C, G, or T, or if more than 1
   alternate base is present, they must be listed, and comma-separated, and there can be no more than 3 alternate bases.}
}
Any columns other than \code{CHROM}, \code{POS}, \code{REF} and \code{ALT} will be ignored.
If consSeq is NULL then V must also have a LENGTH column that gives the number of bases in the
contig.}

\item{targets}{A data frame of the variants that are targets for assay development.  This
can also have whatever columns that are desired, but it must have CHROM and POS, concordant with \code{V},
and it cannot have a column LENGTH in it.
 The extra columns in this
 data frame will all be represented in the output.}

\item{consSeq}{A data frame of consensus sequences.  Must have one column \code{CHROM}, exactly concordant with the
naming conventions of \code{V} and \code{targets}, and one column \code{Seq} which holds the consensus sequences as
strings.  There can be more columns, but they will be ignored.  If this is NULL, then the function returns just the
flanking information summaries, without the sequence from which to build assays. However, in that
case, there must be a LENGTH column in V.}

\item{reqDist}{The required minimum number of bases between a target SNP and the nearest flanking SNP
for a target to be designable (according to whomever's criterion).}

\item{allVar}{Logical indicating whether all variable sites within the CHROMs containing the targets
should be returned, or just the targets themselves.}
}
\value{
Returns a data frame with rows as in \code{targets} with additional columns appended to
it, namely:
\itemize{
   \item One here
   \item One there
}
}
\description{
More later
}

