% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{grab_vcf}
\alias{grab_vcf}
\title{read in a (possibly gzipped) VCF file and return it as a data frame}
\usage{
grab_vcf(path, all_cols = FALSE, top_test = 200)
}
\arguments{
\item{path}{The path to a VCF file (which may or may not be gzipped)}

\item{all_cols}{Logical indicating whether to return all columns (TRUE)
or only CHROM, POS, REF, and ALT.}

\item{top_test}{How many lines to scan from the top of the file to look
for the "#CHROM" showing that it is the header.}
}
\description{
This scans a VCF file to find the header line (by searching for a line
starting with "#CHROM"), and then it reads it in and returns it as a
data frame.  It doesn' worry about the formatting of the genotype columns
because we are typically concerned here just with getting CHROM, POS,
REF, and ALT.  In fact, by default, those four columns are all that
this function returns
}
\examples{
grab_vcf(system.file("textdata", "vcf.txt.gz", package = "snps2assays"))
}

